+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Features:
 - Add Greece 🇬🇷
 - Add Hungary 🇭🇺
 - Add Ireland 🇮🇪
+- Add #normalized to Austria 🇦🇹
 
 ## [0.1.0] - 2025-04-14